requirements.txt
transformers>=4.38.0
trl>=0.8.0
gensim>=4.3.0
datasets>=2.18.0
torch>=2.1.0
anthropic>=0.25.0
wandb>=0.16.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
pyyaml>=6.0.1
accelerate>=0.27.0
sentencepiece>=0.2.0
protobuf>=4.25.0
